[WIP] - refactoring to clean arch
