+++ MyMusic/logs/token-provider.log
@@ -202,3 +202,42 @@ Caused by: java.net.BindException: Endereço já em uso
 2022-08-02 11:18:37,554 INFO org.springframework.orm.jpa.AbstractEntityManagerFactoryBean [SpringApplicationShutdownHook] Closing JPA EntityManagerFactory for persistence unit 'default'
 2022-08-02 11:18:37,557 INFO com.zaxxer.hikari.HikariDataSource [SpringApplicationShutdownHook] HikariPool-1 - Shutdown initiated...
 2022-08-02 11:18:37,559 INFO com.zaxxer.hikari.HikariDataSource [SpringApplicationShutdownHook] HikariPool-1 - Shutdown completed.
+2022-08-02 15:31:16,153 INFO org.springframework.boot.StartupInfoLogger [main] Starting TokenProviderApplication v0.0.1-SNAPSHOT using Java 17.0.3 on lnb036119cps with PID 64241 (/home/cassiaav/IdeaProjects/MyMusic/MyMusic/token-provider-0.0.1-SNAPSHOT.jar started by CIANDT\cassiaav in /home/cassiaav/IdeaProjects/MyMusic/MyMusic)
+2022-08-02 15:31:16,156 INFO org.springframework.boot.SpringApplication [main] No active profile set, falling back to default profiles: default
+2022-08-02 15:31:16,819 INFO org.springframework.data.repository.config.RepositoryConfigurationDelegate [main] Bootstrapping Spring Data JPA repositories in DEFAULT mode.
+2022-08-02 15:31:16,872 INFO org.springframework.data.repository.config.RepositoryConfigurationDelegate [main] Finished Spring Data repository scanning in 45 ms. Found 1 JPA repository interfaces.
+2022-08-02 15:31:17,302 INFO org.springframework.boot.web.embedded.tomcat.TomcatWebServer [main] Tomcat initialized with port(s): 8080 (http)
+2022-08-02 15:31:17,312 INFO org.apache.juli.logging.DirectJDKLog [main] Initializing ProtocolHandler ["http-nio-8080"]
+2022-08-02 15:31:17,313 INFO org.apache.juli.logging.DirectJDKLog [main] Starting service [Tomcat]
+2022-08-02 15:31:17,313 INFO org.apache.juli.logging.DirectJDKLog [main] Starting Servlet engine: [Apache Tomcat/9.0.56]
+2022-08-02 15:31:17,362 INFO org.apache.juli.logging.DirectJDKLog [main] Initializing Spring embedded WebApplicationContext
+2022-08-02 15:31:17,362 INFO org.springframework.boot.web.servlet.context.ServletWebServerApplicationContext [main] Root WebApplicationContext: initialization completed in 1155 ms
+2022-08-02 15:31:17,449 INFO com.zaxxer.hikari.HikariDataSource [main] HikariPool-1 - Starting...
+2022-08-02 15:31:17,598 INFO com.zaxxer.hikari.HikariDataSource [main] HikariPool-1 - Start completed.
+2022-08-02 15:31:17,608 INFO org.springframework.boot.autoconfigure.h2.H2ConsoleAutoConfiguration [main] H2 console available at '/h2-console'. Database available at 'jdbc:h2:mem:testdb'
+2022-08-02 15:31:17,732 INFO org.hibernate.jpa.internal.util.LogHelper [main] HHH000204: Processing PersistenceUnitInfo [name: default]
+2022-08-02 15:31:17,791 INFO org.hibernate.Version [main] HHH000412: Hibernate ORM core version 5.6.4.Final
+2022-08-02 15:31:17,930 INFO org.hibernate.annotations.common.reflection.java.JavaReflectionManager [main] HCANN000001: Hibernate Commons Annotations {5.1.2.Final}
+2022-08-02 15:31:18,031 INFO org.hibernate.dialect.Dialect [main] HHH000400: Using dialect: org.hibernate.dialect.H2Dialect
+2022-08-02 15:31:18,512 INFO org.hibernate.engine.transaction.jta.platform.internal.JtaPlatformInitiator [main] HHH000490: Using JtaPlatform implementation: [org.hibernate.engine.transaction.jta.platform.internal.NoJtaPlatform]
+2022-08-02 15:31:18,521 INFO org.springframework.orm.jpa.AbstractEntityManagerFactoryBean [main] Initialized JPA EntityManagerFactory for persistence unit 'default'
+2022-08-02 15:31:18,558 WARN org.springframework.boot.autoconfigure.orm.jpa.JpaBaseConfiguration$JpaWebConfiguration [main] spring.jpa.open-in-view is enabled by default. Therefore, database queries may be performed during view rendering. Explicitly configure spring.jpa.open-in-view to disable this warning
+2022-08-02 15:31:19,120 INFO org.springframework.boot.actuate.endpoint.web.EndpointLinksResolver [main] Exposing 1 endpoint(s) beneath base path '/actuator'
+2022-08-02 15:31:19,136 INFO org.apache.juli.logging.DirectJDKLog [main] Starting ProtocolHandler ["http-nio-8080"]
+2022-08-02 15:31:19,147 INFO org.springframework.boot.web.embedded.tomcat.TomcatWebServer [main] Tomcat started on port(s): 8080 (http) with context path ''
+2022-08-02 15:31:19,158 INFO org.springframework.boot.StartupInfoLogger [main] Started TokenProviderApplication in 3.33 seconds (JVM running for 3.775)
+2022-08-02 15:35:04,175 INFO org.apache.juli.logging.DirectJDKLog [http-nio-8080-exec-1] Initializing Spring DispatcherServlet 'dispatcherServlet'
+2022-08-02 15:35:04,175 INFO org.springframework.web.servlet.FrameworkServlet [http-nio-8080-exec-1] Initializing Servlet 'dispatcherServlet'
+2022-08-02 15:35:04,177 INFO org.springframework.web.servlet.FrameworkServlet [http-nio-8080-exec-1] Completed initialization in 2 ms
+2022-08-02 15:35:04,281 INFO com.ciandt.token.provider.controllers.TokenController [http-nio-8080-exec-1] Recebido requisição para criação de token para chave: Cassia
+2022-08-02 15:35:04,354 INFO com.ciandt.token.provider.controllers.TokenController [http-nio-8080-exec-1] Token gerado com sucesso
+2022-08-02 15:35:04,377 INFO com.ciandt.token.provider.controllers.TokenAuthorizerController [http-nio-8080-exec-2] Recebido requisição para geração de token: Cassia
+2022-08-02 15:35:04,394 INFO com.ciandt.token.provider.controllers.TokenAuthorizerController [http-nio-8080-exec-2] Token validado com sucesso
+2022-08-02 15:36:21,292 INFO com.ciandt.token.provider.controllers.TokenController [http-nio-8080-exec-4] Recebido requisição para criação de token para chave: Cassia
+2022-08-02 15:36:21,314 INFO com.ciandt.token.provider.controllers.TokenController [http-nio-8080-exec-4] Token gerado com sucesso
+2022-08-02 15:36:21,320 INFO com.ciandt.token.provider.controllers.TokenAuthorizerController [http-nio-8080-exec-5] Recebido requisição para geração de token: Cassia
+2022-08-02 15:36:21,325 INFO com.ciandt.token.provider.controllers.TokenAuthorizerController [http-nio-8080-exec-5] Token validado com sucesso
+2022-08-02 15:41:41,779 INFO com.ciandt.token.provider.controllers.TokenController [http-nio-8080-exec-7] Recebido requisição para criação de token para chave: Cassia
+2022-08-02 15:41:41,784 INFO com.ciandt.token.provider.controllers.TokenController [http-nio-8080-exec-7] Token gerado com sucesso
+2022-08-02 15:41:41,799 INFO com.ciandt.token.provider.controllers.TokenAuthorizerController [http-nio-8080-exec-8] Recebido requisição para geração de token: Cassia
+2022-08-02 15:41:41,802 INFO com.ciandt.token.provider.controllers.TokenAuthorizerController [http-nio-8080-exec-8] Token validado com sucesso
