+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/adapter/controller/dto/ResponseWrapper.java
@@ -1,0 +1,15 @@
+package com.ciandt.summit.bootcamp2022.adapter.controller.dto;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class ResponseWrapper<T> {
+
+  private T data;
+
+  public ResponseWrapper(T data) {
+    this.data = data;
+  }
+}
