+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/adapter/controller/validation/ValidationErrorDto.java
@@ -1,0 +1,15 @@
+package com.ciandt.summit.bootcamp2022.adapter.controller.validation;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class ValidationErrorDto {
+
+  private String field;
+  private String error;
+
+}
