+++ MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/adapter/controller/validation/ValidationErrorHandler.java
@@ -1,0 +1,37 @@
+package com.ciandt.summit.bootcamp2022.adapter.controller.validation;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
+import org.springframework.context.i18n.LocaleContextHolder;
+import org.springframework.http.HttpStatus;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@RestControllerAdvice
+public class ValidationErrorHandler {
+
+  @Autowired
+  private MessageSource messageSource;
+
+  @ResponseStatus(code = HttpStatus.BAD_REQUEST)
+  @ExceptionHandler(MethodArgumentNotValidException.class)
+  public List<ValidationErrorDto> handle(MethodArgumentNotValidException exception) {
+    List<ValidationErrorDto> dto = new ArrayList<>();
+
+    List<FieldError> fieldErrors = exception.getBindingResult().getFieldErrors();
+    fieldErrors.forEach(e -> {
+      String mensagem = messageSource.getMessage(e, LocaleContextHolder.getLocale());
+      ValidationErrorDto erro = new ValidationErrorDto(e.getField(), mensagem);
+      dto.add(erro);
+    });
+
+    return dto;
+  }
+
+}
