--- MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/controller/CacheController.java
@@ -1,19 +1,0 @@
-package com.ciandt.summit.bootcamp2022.controller;
-
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/cache")
-public class CacheController {
-
-    @GetMapping("/aW52YWxpZGEgcmVsYXTDs3JpbyBkZSB2ZW5kYXM")
-    @CacheEvict(value = "listOfMusicByNameOrArtist", allEntries = true)
-    public ResponseEntity<?> invalidateMusicCache() {
-        return ResponseEntity.ok().build();
-    }
-
-}
