--- MyMusic/src/main/java/com/ciandt/summit/bootcamp2022/controller/dto/ResponseWrapper.java
@@ -1,15 +1,0 @@
-package com.ciandt.summit.bootcamp2022.controller.dto;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class ResponseWrapper<T> {
-
-  private T data;
-
-  public ResponseWrapper(T data) {
-    this.data = data;
-  }
-}
